+++ alb.tf
@@ -1,21 +1,19 @@
-resource "aws_alb" "alb_jenkins_master" {
-  name                       = "alb-jenkins-master"
+resource "aws_alb" "alb_jenkins_controller" {
+  name                       = "alb-jenkins-controller"
   internal                   = false
   load_balancer_type         = "application"
   security_groups            = [aws_security_group.alb_security_group.id]
   subnets                    = var.public_subnets
   ip_address_type            = "ipv4"
   enable_deletion_protection = false
-  tags                       = var.default_tags
 }
 
-resource "aws_alb_target_group" "jenkins_master_tg" {
-  name        = "alb-http-jenkins-master"
-  port        = var.master_listening_port
+resource "aws_alb_target_group" "jenkins_controller_tg" {
+  name        = "alb-http-jenkins-controller"
+  port        = var.controller_listening_port
   target_type = "ip"
   protocol    = "HTTP"
   vpc_id      = var.vpc_id
-  tags        = var.default_tags
 
   stickiness {
     type    = "lb_cookie"
@@ -36,22 +34,22 @@ resource "aws_alb_target_group" "jenkins_master_tg" {
 }
 
 # This listener is used when dont't use https with the ALB
-resource "aws_lb_listener" "master_http" {
+resource "aws_lb_listener" "controller_http" {
   count             = var.route53_zone_name != "" ? 0 : 1
-  load_balancer_arn = aws_alb.alb_jenkins_master.arn
+  load_balancer_arn = aws_alb.alb_jenkins_controller.arn
   port              = "80"
   protocol          = "HTTP"
 
   default_action {
     type             = "forward"
-    target_group_arn = aws_alb_target_group.jenkins_master_tg.arn
+    target_group_arn = aws_alb_target_group.jenkins_controller_tg.arn
   }
 }
 
 
-resource "aws_lb_listener" "master_http_redirect" {
+resource "aws_lb_listener" "controller_http_redirect" {
   count             = var.route53_zone_name != "" ? 1 : 0
-  load_balancer_arn = aws_alb.alb_jenkins_master.arn
+  load_balancer_arn = aws_alb.alb_jenkins_controller.arn
   port              = "80"
   protocol          = "HTTP"
 
@@ -66,16 +64,16 @@ resource "aws_lb_listener" "master_http_redirect" {
   }
 }
 
-resource "aws_lb_listener" "master_https" {
+resource "aws_lb_listener" "controller_https" {
   count             = var.route53_zone_name != "" ? 1 : 0
-  load_balancer_arn = aws_alb.alb_jenkins_master.arn
+  load_balancer_arn = aws_alb.alb_jenkins_controller.arn
   port              = "443"
   protocol          = "HTTPS"
   ssl_policy        = "ELBSecurityPolicy-TLS-1-2-Ext-2018-06"
-  certificate_arn   = aws_acm_certificate.master_certificate.0.arn
+  certificate_arn   = aws_acm_certificate.controller_certificate[0].arn
 
   default_action {
     type             = "forward"
-    target_group_arn = aws_alb_target_group.jenkins_master_tg.arn
+    target_group_arn = aws_alb_target_group.jenkins_controller_tg.arn
   }
 }