+++ security-groups.tf
@@ -1,15 +1,15 @@
-resource "aws_security_group" "jenkins_master_sg" {
-  name        = "sgr-jenkins-master-service"
-  description = "Jenkins Master ECS service security group."
+resource "aws_security_group" "jenkins_controller_ecs_service" {
+  name        = "sgr-jenkins-controller-ecs-service"
+  description = "Jenkins Controller ECS service security group."
   vpc_id      = var.vpc_id
-  tags        = merge({ "Name" : "sgr-jenkins-master-service" }, var.default_tags)
+  tags        = { "Name" : "sgr-jenkins-controller-service" }
 }
 
 resource "aws_security_group" "alb_security_group" {
-  name        = "sgr-jenkins-master-alb"
-  description = "Jenkins Master ALB security group."
+  name        = "sgr-jenkins-controller-alb"
+  description = "Jenkins Controller ALB security group."
   vpc_id      = var.vpc_id
-  tags        = merge({ "Name" : "sgr-jenkins-master-alb" }, var.default_tags)
+  tags        = { "Name" : "sgr-jenkins-controller-alb" }
 }
 
 resource "aws_security_group_rule" "alb_ingress_http" {
@@ -18,7 +18,7 @@ resource "aws_security_group_rule" "alb_ingress_http" {
   to_port           = 80
   protocol          = "tcp"
   type              = "ingress"
-  cidr_blocks       = ["0.0.0.0/0"]
+  cidr_blocks       = var.allowed_ip_addresses
 }
 
 resource "aws_security_group_rule" "alb_ingress_https" {
@@ -28,7 +28,7 @@ resource "aws_security_group_rule" "alb_ingress_https" {
   to_port           = 443
   protocol          = "tcp"
   type              = "ingress"
-  cidr_blocks       = ["0.0.0.0/0"]
+  cidr_blocks       = var.allowed_ip_addresses
 }
 
 resource "aws_security_group_rule" "alb_egress_all" {
@@ -40,41 +40,41 @@ resource "aws_security_group_rule" "alb_egress_all" {
   cidr_blocks       = ["0.0.0.0/0"]
 }
 
-resource "aws_security_group_rule" "jenkins_master_ingress_alb" {
-  security_group_id        = aws_security_group.jenkins_master_sg.id
-  from_port                = var.master_listening_port
-  to_port                  = var.master_listening_port
+resource "aws_security_group_rule" "jenkins_controller_ingress_alb" {
+  security_group_id        = aws_security_group.jenkins_controller_ecs_service.id
+  from_port                = var.controller_listening_port
+  to_port                  = var.controller_listening_port
   protocol                 = "tcp"
   type                     = "ingress"
   source_security_group_id = aws_security_group.alb_security_group.id
-  description              = "From ALB to Jenkins Master listening port."
+  description              = "From ALB to Jenkins Controller listening port."
 }
 
 resource "aws_security_group_rule" "allow_agents_to_jks_jnlp_port" {
-  count             = length(var.private_subnets)
-  security_group_id = aws_security_group.jenkins_master_sg.id
-  from_port         = var.master_jnlp_port
-  to_port           = var.master_jnlp_port
+  for_each          = var.private_subnets
+  security_group_id = aws_security_group.jenkins_controller_ecs_service.id
+  from_port         = var.controller_jnlp_port
+  to_port           = var.controller_jnlp_port
   protocol          = "tcp"
   type              = "ingress"
-  cidr_blocks       = list("${data.aws_network_interface.private_nlb_network_interface[count.index].private_ip}/32")
-  description       = "From NLB to Jenkins Master JNLP via ENI ${data.aws_network_interface.private_nlb_network_interface[count.index].id}."
+  cidr_blocks       = ["${data.aws_network_interface.each_network_interface[each.key].private_ip}/32"]
+  description       = "From the NLB to the Jenkins Controller via JNLP and ENI ${data.aws_network_interface.each_network_interface[each.key].id}."
 }
 
 # When using a private nlb we need to have this rule for nlb health check to work.
 resource "aws_security_group_rule" "from_private_nlb_network_interfaces" {
-  count             = length(var.private_subnets)
-  security_group_id = aws_security_group.jenkins_master_sg.id
-  from_port         = var.master_listening_port
-  to_port           = var.master_listening_port
+  for_each          = var.private_subnets
+  security_group_id = aws_security_group.jenkins_controller_ecs_service.id
+  from_port         = var.controller_listening_port
+  to_port           = var.controller_listening_port
   protocol          = "tcp"
   type              = "ingress"
-  cidr_blocks       = list("${data.aws_network_interface.private_nlb_network_interface[count.index].private_ip}/32")
-  description       = "From NLB to Jenkins Master HTTP via ENI ${data.aws_network_interface.private_nlb_network_interface[count.index].id}. Required for health check."
+  cidr_blocks       = ["${data.aws_network_interface.each_network_interface[each.key].private_ip}/32"]
+  description       = "From the NLB to the Jenkins Controller via HTTP and ENI ${data.aws_network_interface.each_network_interface[each.key].id}. Required for health check."
 }
 
-resource "aws_security_group_rule" "master_egress_all" {
-  security_group_id = aws_security_group.jenkins_master_sg.id
+resource "aws_security_group_rule" "controller_egress_all" {
+  security_group_id = aws_security_group.jenkins_controller_ecs_service.id
   from_port         = 0
   to_port           = 0
   protocol          = "-1"
@@ -87,7 +87,7 @@ resource "aws_security_group" "jenkins_agents" {
   name        = "sgr-jenkins-agents"
   description = "Security group attached to Jenkins agents running in Fargate."
   vpc_id      = var.vpc_id
-  tags        = merge({ "Name" : "sgr-jenkins-agents" }, var.default_tags)
+  tags        = { "Name" : "sgr-jenkins-agents" }
 }
 
 resource "aws_security_group_rule" "jenkins_agent_egress" {
@@ -101,10 +101,10 @@ resource "aws_security_group_rule" "jenkins_agent_egress" {
 
 ### EFS
 resource "aws_security_group" "efs" {
-  name        = "sgr-jenkins-master-efs"
-  description = "Jenkins Master EFS security group."
+  name        = "sgr-jenkins-controller-efs"
+  description = "Jenkins Controller EFS security group."
   vpc_id      = var.vpc_id
-  tags        = merge({ "Name" : "sgr-jenkins-master-efs" }, var.default_tags)
+  tags        = { "Name" : "sgr-jenkins-controller-efs" }
 }
 
 resource "aws_security_group_rule" "allow_jenkins_to_efs" {
@@ -113,6 +113,6 @@ resource "aws_security_group_rule" "allow_jenkins_to_efs" {
   to_port                  = 2049
   protocol                 = "tcp"
   type                     = "ingress"
-  description              = "Jenkins Master access to EFS."
-  source_security_group_id = aws_security_group.jenkins_master_sg.id
+  description              = "Jenkins Controller access to EFS."
+  source_security_group_id = aws_security_group.jenkins_controller_ecs_service.id
 }