+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # v1.0.1
- - fix: Add support for AWS provider V3
- - fix: Use variable efs_provisioned_throughput_in_mibps
+
+- fix: Add support for AWS provider V3
+- fix: Use variable efs_provisioned_throughput_in_mibps
 
 # v1.0.0
+
 Initial version